Update magic value to PULSE in hex

+++ target_chains/ethereum/contracts/forge-test/PulseSchedulerGovernance.t.sol
@@ -8,7 +8,7 @@ import "@openzeppelin/contracts/proxy/ERC1967/ERC1967Proxy.sol";
 import "../contracts/pulse/SchedulerUpgradeable.sol";
 
 contract InvalidMagicPulseScheduler is SchedulerUpgradeable {
-    function schedulerUpgradableMagic() public pure override returns (uint32) {
+    function schedulerUpgradableMagic() public pure override returns (uint64) {
         return 0x12345678; // Different magic
     }
 }
